Fix error when Authorization header is blank
If a request is made to a bearer token resource with a blank
(but present) Authorization: header, rack-oauth2 will cause
rack to throw a "undefined method `downcase' for nil:NilClass"
error. This checks for a nil value and treats the request as if
the header was not provided, preventing the error.